requirements.txt
numpy==1.21.2
pandas==1.3.3
psycopg2==2.9.1
openai==0.27.0
python-dotenv==0.19.1

